Create endpoint to shopping list

+++ src/main/java/com/restaurant/app/food/controller/dto/FoodRequest.java
@@ -3,6 +3,6 @@ package com.restaurant.app.food.controller.dto;
 public record FoodRequest(
         Long positionId,
         String foodName,
-        Integer foodPrice
+        Double foodPrice
 ) {
 }
