+++ src/main/java/com/restaurant/app/food/controller/dto/FoodRequestResponse.java
@@ -5,6 +5,6 @@ public record FoodRequestResponse(
         Long categoryId,
         Long positionId,
         String foodName,
-        Integer foodPrice
+        Double foodPrice
 ) {
 }
