+++ src/main/java/com/restaurant/app/food/repository/dto/FoodModel.java
@@ -5,6 +5,6 @@ public record FoodModel(
         Long categoryId,
         Long positionId,
         String foodName,
-        Integer foodPrice
+        Double foodPrice
 ) {
 }
