+++ src/main/java/com/restaurant/app/user/controller/dto/UserRequestResponse.java
@@ -10,7 +10,7 @@ public record UserRequestResponse(
         String role,
         String address,
         String phoneNumber,
-        Long money,
+        Double money,
         Boolean loyaltyCard
 ) {
 }
